Update the gui
a better gui buttons control
make the main more clean
