+++ source/utils.hpp
@@ -1,0 +1,63 @@
+void close()
+{
+	//Free loaded images
+	gTextTexture.free();
+	Farest.free();
+	Heart.free();
+	TPreview.free();
+	TChapters.free();
+	TSearchPreview.free();
+	//Free global font
+	TTF_CloseFont(gFont);
+	gFont = NULL;
+	TTF_CloseFont(digifont);
+	digifont = NULL;
+
+	TTF_CloseFont(gFontcapit);
+	gFontcapit = NULL;
+	TTF_CloseFont(gFont2);
+	gFont2 = NULL;
+	TTF_CloseFont(gFont3);
+	gFont3 = NULL;
+	//Destroy window	
+	SDL_DestroyRenderer(gRenderer);
+	SDL_DestroyWindow(gWindow);
+	gWindow = NULL;
+	gRenderer = NULL;
+	//Free the music
+	Mix_FreeMusic(gMusic);
+	gMusic = NULL;
+	Mix_Quit();
+	//Quit SDL subsystems
+	TTF_Quit();
+	IMG_Quit();
+	SDL_Quit();
+}
+
+void replace(std::string& subject, const std::string& search,
+	const std::string& replace) {
+	size_t pos = 0;
+	while ((pos = subject.find(search, pos)) != std::string::npos) {
+		subject.replace(pos, search.length(), replace);
+		pos += replace.length();
+	}
+}
+
+
+void mayus(std::string &s)
+{
+	bool cap = true;
+
+	for (unsigned int i = 0; i <= s.length(); i++)
+	{
+		if (isalpha(s[i]) && cap == true)
+		{
+			s[i] = toupper(s[i]);
+			cap = false;
+		}
+		else if (isspace(s[i]))
+		{
+			cap = true;
+		}
+	}
+}
